- Now requiring reauthentication if on canceled (deleted) device. - Streamlined reauthentication flow into single signOut method on   OTableViewController. - Additional switchboard actions in member view:   + Calling residence home numbers.   + Interacting with guardians of parentless minors. - Replaced add button with action button in member view. Added actions for:   + Adding address.   + Changing password (if user).   + Editing relations (if ward).   + Correcting gender (jury's still out on this one). - Performance improvements.   + Using single static NSCalendar instance for calendar operations.   + Replaced sharedDatePicker instance with view specific instances, but     only instantiating if view is editable, and via viewDidAppear to avoid     delay. - Using token to avoid concurrent replication. - General code streamlining. - Fixed race condition when running on device. - Other bugfixes.

+++ OrigoApp/OAlert.h
@@ -13,8 +13,8 @@
 + (void)showAlertWithTitle:(NSString *)title text:(NSString *)text;
 + (void)showAlertWithTitle:(NSString *)title text:(NSString *)text tag:(NSInteger)tag;
 + (void)showAlertForError:(NSError *)error;
-+ (void)showAlertForError:(NSError *)error tag:(int)tag delegate:(id)delegate;
++ (void)showAlertForError:(NSError *)error tag:(NSInteger)tag delegate:(id)delegate;
 + (void)showAlertForHTTPStatus:(NSInteger)status;
-+ (void)showAlertForHTTPStatus:(NSInteger)status tag:(int)tag delegate:(id)delegate;
++ (void)showAlertForHTTPStatus:(NSInteger)status tag:(NSInteger)tag delegate:(id)delegate;
 
 @end
